requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
lxml==5.1.0
jinja2==3.1.3
